+++ WooCommerce/Legacy/Tax.cs
@@ -10,7 +10,7 @@ namespace WooCommerceNET.WooCommerce.Legacy
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
         /// <summary>
         /// Country code. See ISO 3166 Codes (Countries) for more details
