+++ WooCommerce/Legacy/Webhook.cs
@@ -19,7 +19,7 @@ namespace WooCommerceNET.WooCommerce.Legacy
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
 
         /// <summary>
@@ -111,7 +111,7 @@ namespace WooCommerceNET.WooCommerce.Legacy
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
         /// <summary>
         /// The delivery duration, in seconds        
@@ -221,13 +221,13 @@ namespace WooCommerceNET.WooCommerce.Legacy
         /// The webhook’s ID
         /// </summary>
         [DataMember(Name = "X-WC-Webhook-ID", EmitDefaultValue = false)]
-        public int XWCWebhookID { get; set; }
+        public uint XWCWebhookID { get; set; }
 
         /// <summary>
         /// The delivery ID
         /// </summary>
         [DataMember(Name = "X-WC-Webhook-Delivery-ID", EmitDefaultValue = false)]
-        public int XWCWebhookDeliveryID { get; set; }
+        public uint XWCWebhookDeliveryID { get; set; }
 
     }
