+++ WooCommerce/v1/Customer.cs
@@ -112,7 +112,7 @@ namespace WooCommerceNET.WooCommerce.v1
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
         /// <summary>
         /// UTC DateTime of the customer last order.
