+++ WooCommerce/v2/Variation.cs
@@ -15,7 +15,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
         /// <summary>
         /// The date the variation was created, in the site’s timezone. 
@@ -240,7 +240,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// Menu order, used to custom sort products.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int menu_order { get; set; }
+        public uint menu_order { get; set; }
 
         /// <summary>
         /// Meta data. See Product variation - Meta data properties
@@ -309,7 +309,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// Image ID.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
         /// <summary>
         /// The date the image was created, in the site’s timezone. 
@@ -371,7 +371,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// Attribute ID.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
         /// <summary>
         /// Attribute name.
