+++ WooCommerce/v3/Webhook.cs
@@ -15,7 +15,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public uint? id { get; set; }
+        public ulong? id { get; set; }
 
         /// <summary>
         /// The delivery duration, in seconds.     
@@ -129,13 +129,13 @@ namespace WooCommerceNET.WooCommerce.v3
         /// The webhook’s ID
         /// </summary>
         [DataMember(Name = "X-WC-Webhook-ID", EmitDefaultValue = false)]
-        public int XWCWebhookID { get; set; }
+        public uint XWCWebhookID { get; set; }
 
         /// <summary>
         /// The delivery ID
         /// </summary>
         [DataMember(Name = "X-WC-Webhook-Delivery-ID", EmitDefaultValue = false)]
-        public int XWCWebhookDeliveryID { get; set; }
+        public uint XWCWebhookDeliveryID { get; set; }
 
     }
